--- gemini-core/src/main/java/com/haroncode/gemini/connection/util/Mapper.kt
@@ -1,6 +1,0 @@
-package com.haroncode.gemini.connection.util
-
-/**
- * @author HaronCode.
- */
-typealias Mapper<In, Out> = (In) -> Out
