+++ gemini-dsl/src/main/java/com/haroncode/gemini/dsl/BaseConnectionRuleDsl.kt
@@ -26,8 +26,8 @@ inline infix fun <T : Any, R : Any> Publisher<T>.connectTo(consumer: Consumer<R>
 inline infix fun <Out : Any, In : Any> BaseConnectionRule<Out, In>.with(
     crossinline transformer: Transformer<In, In>
 ): BaseConnectionRule<Out, In> = BaseConnectionRule(
-    consumer = this.consumer,
-    publisher = this.publisher,
+    consumer = consumer,
+    publisher = publisher,
     transformer = FlowableTransformer { stream ->
         Flowable.fromPublisher(this.transformer.apply(stream))
             .compose { input -> transformer.invoke(input) }
@@ -37,24 +37,24 @@ inline infix fun <Out : Any, In : Any> BaseConnectionRule<Out, In>.with(
 inline infix fun <Out : Any, In : Any> Pair<Publisher<Out>, Consumer<In>>.with(
     crossinline transformer: Transformer<Out, In>
 ) = BaseConnectionRule(
-    consumer = this.second,
-    publisher = this.first,
+    consumer = second,
+    publisher = first,
     transformer = FlowableTransformer { stream -> transformer.invoke(stream) }
 )
 
 inline infix fun <Out : Any, In : Any> Pair<Publisher<Out>, Consumer<In>>.mapper(
     crossinline mapper: Mapper<Out, In>
 ) = BaseConnectionRule(
-    consumer = this.second,
-    publisher = this.first,
+    consumer = second,
+    publisher = first,
     transformer = FlowableTransformer { stream -> stream.map { input -> mapper.invoke(input) } }
 )
 
 inline infix fun <T : Any, R : Any> BaseConnectionRule<T, R>.observeOn(
     scheduler: Scheduler
 ): BaseConnectionRule<T, R> = BaseConnectionRule(
-    consumer = this.consumer,
-    publisher = this.publisher,
+    consumer = consumer,
+    publisher = publisher,
     transformer = FlowableTransformer { stream ->
         Flowable.fromPublisher(this.transformer.apply(stream))
             .observeOn(scheduler)
@@ -64,6 +64,6 @@ inline infix fun <T : Any, R : Any> BaseConnectionRule<T, R>.observeOn(
 inline infix fun <Event : Any> Store<*, *, Event>.eventsTo(
     eventListener: EventListener<Event>
 ) = EventListenerConnection(
-    eventPublisher = this.eventSource,
+    eventPublisher = eventSource,
     eventListener = eventListener
 )
