+++ gemini-dsl/src/main/java/com/haroncode/gemini/dsl/Mapper.kt
@@ -1,0 +1,6 @@
+package com.haroncode.gemini.dsl
+
+/**
+ * @author HaronCode.
+ */
+typealias Mapper<In, Out> = (In) -> Out
